pyproject.toml
[project]
name = "weather-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=3.0.6",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
]
